Override Strongback's incorrect test classpath.
Override Strongback's incorrect buildpath for test compilation. Why?
Strongback assumes that the correct class path is ${classpath}, which
does have WPILib.jar and NetworkTables.jar. However, this doesn't include
any user libraries (like Strongback itself!). Instead we add the path "classpath.path",
which is the correct path that wpilib uses for compilation and includes user libraries.
(the ${classpath} property is defined in ${wpilib.dir}/java/current/ant/build.properties.)

TestRobot.java has been updated to test these changes.
