--- testsrc/org/usfirst/frc/team1076/robot/TestRobot.java
@@ -1,19 +1,0 @@
-package org.usfirst.frc.team1076.robot;
-
-import static org.junit.Assert.*;
-
-import org.junit.Test;
-
-public class TestRobot {
-
-    @Test
-    public void test() {
-        assertTrue(true);
-    }
-    
-    @Test(expected = AssertionError.class)
-    public void testFail() {
-        assertTrue(false);
-    }
-
-}
